+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Karafka Web changelog
 
 ## 0.1.4 (Unreleased)
-- Make sure that states and reports are always dispatched to the partition `0`. This should prevent UI from not fully working when someone accidentally creates more partitions than expected.
-- Fix a bug where bootstrapping would create two initial states.
-- Introduce a /status page that can validate the setup and tell what is missing (#1318)
+- [Feature] Introduce a /status page that can validate the setup and tell what is missing (#1318)
+- [Improvement] Allow for disabling the consumer subscription from Web for multi-tenant Web UI usage (#1331)
+- [Improvement] Make sure that states and reports are always dispatched to the partition `0`. This should prevent UI from not fully working when someone accidentally creates more partitions than expected.
+- [Fix] Fix a bug where bootstrapping would create two initial states.
 
 ## 0.1.3 (2023-02-14)
 - Skip topics creation if web topics already exist (do not raise error)