+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Karafka Web changelog
 
-## 0.1.4 (Unreleased)
-- [Feature] Introduce a /status page that can validate the setup and tell what is missing (#1318)
+## 0.2.0 (Unreleased)
+- **[Feature]** Introduce ability to tag `Karafka::Process` to display process-centric tags in the Web UI.
+- **[Feature]** Introduce ability to tag consumer instances to display consumption-centric tags in the Web UI.
+- **[Feature]** Introduce a /status page that can validate the setup and tell what is missing (#1318)
 - [Improvement] Allow for disabling the consumer subscription from Web for multi-tenant Web UI usage (#1331)
 - [Improvement] Make sure that states and reports are always dispatched to the partition `0`. This should prevent UI from not fully working when someone accidentally creates more partitions than expected.
 - [Fix] Fix a bug where bootstrapping would create two initial states.